+++ lukefi/metsi/data/model.py
@@ -340,8 +340,8 @@ class ForestStand(Finalizable, ComputationalUnit):
     reference_trees_pre_vec: list[ReferenceTree] = dataclasses.field(default_factory=list)
     tree_strata_pre_vec: list[TreeStratum] = dataclasses.field(default_factory=list)
 
-    reference_trees: ReferenceTrees = ReferenceTrees()
-    tree_strata: TreeStrata = TreeStrata()
+    reference_trees: ReferenceTrees = dataclasses.field(default_factory=ReferenceTrees)
+    tree_strata: TreeStrata = dataclasses.field(default_factory=TreeStrata)
 
     # unique identifier for entity within its domain
     identifier: str = ""